% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulkdata}
\alias{bulkdata}
\title{Whole blood DNA methylation profiles.}
\format{
A matrix with 599 rows and 10 variables
}
\usage{
bulkdata
}
\description{
A sample dataset of whole blood DNA methylation profiles from 10 individuals
}
\keyword{datasets}
