% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBulkProfiles.R
\name{createBulkProfiles}
\alias{createBulkProfiles}
\title{Generate bulk tissue profiles from purified cell types}
\usage{
createBulkProfiles(purifiedBetas, matrixSimProp)
}
\arguments{
\item{purifiedBetas}{}

\item{matrixSimProp}{}
}
\value{

}
\description{
createBulkProfiles
}
