% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCellProportions.R
\name{generateCellProportions}
\alias{generateCellProportions}
\title{Create matrix of cellular composition profiles consisting of proportion
of noise and remaining compostion in the ratio of supplied cell types}
\usage{
generateCellProportions(cellProp, noise)
}
\arguments{
\item{cellProp}{A vector of the required cellular proportions. Must be
named with cell labels.}

\item{noise}{A vector of the proportion of noise.}
}
\value{
A matrix with length(noise) columns and length(cellProp)+1
columns.
}
\description{
Create matrix of cellular composition profiles consisting of proportion
of noise and remaining compostion in the ratio of supplied cell types
}
\examples{
meanBloodProp = c(3.01,13.4, 6.13)
names(meanBloodProp)<-c("Bcell", "CDT4+", "CDT8")
noise = c(seq(0,0.1,0.02))
generateCellProportions(meanBloodProp, noise)
}
