% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelBloodCoef}
\alias{modelBloodCoef}
\title{Coefficients for 6 blood cell types}
\format{
A matrix with 600 rows and 6 variables
}
\usage{
modelBloodCoef
}
\description{
Coefficients for blood cell types for use in deconvoluting cell composition
from whole blood profiles. These data were generated using the
pickCompProbes() function from the minfi package and DNA methylation data
from the FlowSorted.Blood.450k R package.
}
\keyword{datasets}
