% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectCellTypeWithError.R
\name{projectCellTypeWithError}
\alias{projectCellTypeWithError}
\title{Estimate cellular composition with error}
\usage{
projectCellTypeWithError(YIN, coefs)
}
\arguments{
\item{YIN}{a matrix of DNA methylation levels from the samples that require cell
composition to be estimated.}

\item{coefs}{A matrix of deconvolution model parameters including coefficients.
This input can be generated with the pickCompProbesMatrix()$coefs}
}
\value{
A matrix with the estimated proportion of cell types, CETYGO and the number of
sites missing from the model.
}
\description{
Estimate the cellular composition from a DNA methylation profile and calculate
the CETYGO score error metric associated with this estimate.
}
