% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root mean square error (RMSE)}
\usage{
RMSE(m, o)
}
\arguments{
\item{m}{A vector of expected values}

\item{o}{A vector sof observed values expected to be the same length as m}
}
\value{
the root mean square error calculated from the two input vectors
}
\description{
Calculates the root mean squared error from two vectors of observed and expected data.
Assumes the two vectors are of the same length and that the pairs are the same order.
}
\examples{
expected<-c(0.4,0.5,0.2)
observed<-c(0.38, 0.47, 0.25)
}
