% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNoiseSimulations.R
\name{runNoiseSimulations}
\alias{runNoiseSimulations}
\title{Perform simulations to test deconvolution with proportion of noise}
\usage{
runNoiseSimulations(
  trainBetas,
  trainCellTypes,
  trainCellInd,
  testBetas,
  matrixSimProp
)
}
\arguments{
\item{trainBetas}{}

\item{trainCellTypes}{}

\item{trainCellInd}{}

\item{testBetas}{}

\item{matrixSimProp}{}
}
\value{

}
\description{
createBulkProfiles
}
